Decompose parser section state and recovery helpers
